Revert "Test cover"
