+++ src/test/java/com/udhd/apiserver/service/AlbumServiceTest.java
@@ -1,20 +1,36 @@
 package com.udhd.apiserver.service;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.udhd.apiserver.domain.album.AlbumRepository;
+import com.udhd.apiserver.domain.photo.PhotoRepository;
+import com.udhd.apiserver.web.dto.album.AlbumOutlineDto;
 import org.bson.types.ObjectId;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.dao.DuplicateKeyException;
+import org.springframework.test.web.servlet.MockMvc;
 
+import java.sql.Array;
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.mongodb.assertions.Assertions.assertTrue;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.when;
+import static org.mockito.BDDMockito.given;
 
-@ExtendWith(MockitoExtension.class)
+
+@SpringBootTest
 public class AlbumServiceTest {
+    @Mock
+    private PhotoRepository photoRepository;
+
+    @Autowired
+    protected ObjectMapper objectMapper;
 
     @Mock
     AlbumRepository albumRepository;
@@ -22,17 +38,17 @@ public class AlbumServiceTest {
     @InjectMocks
     AlbumService albumService;
 
+    protected MockMvc mockMvc;
+
     @Test
     @DisplayName("중복된 사진 앨범 추가")
-    void saveAlbum_Duplicate() {
+    void saveAlbum_Duplicate() throws Exception {
         String userId = "60e2fea74c17cf5152fb5b78";
         String photoId = "60e2fea74c17cf5152fb5b78";
 
         // 이미 가지고 있는 앨범 데이터
-        when(albumRepository.findByUserIdAndFeedId(new ObjectId(userId), new ObjectId(photoId)))
-                .thenThrow(DuplicateKeyException.class);
-        /*given(albumRepository.findByUserIdAndFeedId(new ObjectId(userId), new ObjectId(photoId)))
-                .willThrow(DuplicateKeyException.class);*/
+        given(albumRepository.findByUserIdAndFeedId(new ObjectId(userId), new ObjectId(photoId)))
+                .willThrow(DuplicateKeyException.class);
 
         assertThrows(DuplicateKeyException.class,
                 () -> albumService.saveAlbum(userId, photoId));
