+++ src/test/java/com/udhd/apiserver/service/PhotoServiceTest.java
@@ -1,5 +1,6 @@
 package com.udhd.apiserver.service;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.udhd.apiserver.domain.album.Album;
 import com.udhd.apiserver.domain.photo.Photo;
 import com.udhd.apiserver.domain.photo.PhotoRepository;
@@ -7,10 +8,10 @@ import com.udhd.apiserver.exception.album.AlbumNotFoundException;
 import com.udhd.apiserver.web.dto.photo.PhotoDetailDto;
 import org.bson.types.ObjectId;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
 
 import java.util.Arrays;
 import java.util.Date;
@@ -21,17 +22,17 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
 
 
-
-@ExtendWith(MockitoExtension.class)
+@SpringBootTest
 public class PhotoServiceTest {
     @Mock
     private PhotoRepository photoRepository;
-    @Mock
-    AlbumService albumService;
 
-    @InjectMocks
-    PhotoService photoService;
+    @Autowired
+    protected ObjectMapper objectMapper;
 
+    @Mock AlbumService albumService;
+
+    protected MockMvc mockMvc;
 
     private final PhotoDetailDto mockPhotoDetailDto
             = PhotoDetailDto.builder()
@@ -45,8 +46,8 @@ public class PhotoServiceTest {
 
 
     @Test
-    void detailPhotoByUserId() {
-        //PhotoService photoService = new PhotoService(albumService, photoRepository);
+    void detailPhotoByUserId() throws Exception {
+        PhotoService photoService = new PhotoService(albumService, photoRepository);
         // given
         String photoId = "6110066423a94f0000000000";
         String photoId_A = "6110066423a94f0000000001";
@@ -73,6 +74,8 @@ public class PhotoServiceTest {
                 .build();
 
         given(albumService.getAlbumDetail(userId, photoId)).willReturn(albums.get(0));
+        given(albumService.getAlbumDetail(userId, photoId_A)).willReturn(albums.get(1));
+
 
         given(photoRepository.findById(new ObjectId(photoId))).willReturn(Optional.ofNullable(photo));
         PhotoDetailDto photoDetailDto = photoService.getPhotoDetail(userId, photoId);
