--- src/test/java/com/udhd/apiserver/web/ControllerTest.java
@@ -1,110 +1,0 @@
-package com.udhd.apiserver.web;
-
-import capital.scalable.restdocs.AutoDocumentation;
-import capital.scalable.restdocs.SnippetRegistry;
-import capital.scalable.restdocs.jackson.JacksonResultHandlers;
-import capital.scalable.restdocs.response.ResponseModifyingPreprocessors;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.udhd.apiserver.util.SecurityUtils;
-import org.junit.Ignore;
-import org.junit.jupiter.api.*;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.MockedStatic;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
-import org.springframework.restdocs.cli.CliDocumentation;
-import org.springframework.restdocs.http.HttpDocumentation;
-import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation;
-import org.springframework.restdocs.operation.preprocess.Preprocessors;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.request.RequestPostProcessor;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-
-import static capital.scalable.restdocs.misc.AuthorizationSnippet.documentAuthorization;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.mockStatic;
-
-@ExtendWith(RestDocumentationExtension.class)
-@AutoConfigureRestDocs
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@SpringBootTest
-@Disabled
-abstract public class ControllerTest {
-
-    protected MockMvc mockMvc;
-    @Autowired
-    protected ObjectMapper objectMapper;
-
-    @Value("${userId}")
-    protected String userId;
-    @Value("${feedId}")
-    protected String feedId;
-
-    @Autowired
-    private WebApplicationContext context;
-
-    private MockedStatic<SecurityUtils> mockedSecurityUtils;
-
-    @BeforeAll
-    public void mockStaticSetup() {
-        mockedSecurityUtils = mockStatic(SecurityUtils.class);
-        given(SecurityUtils.getLoginUserId()).willReturn(userId);
-    }
-
-    @AfterAll
-    public void demockStaticSetup() {
-        mockedSecurityUtils.close();
-    }
-
-    @BeforeEach
-    public void setUp(RestDocumentationContextProvider restDocumentation) throws Exception {
-        this.mockMvc = MockMvcBuilders
-                .webAppContextSetup(context)
-                .alwaysDo(JacksonResultHandlers.prepareJackson(objectMapper))
-                .alwaysDo(MockMvcRestDocumentation.document("{class-name}/{method-name}",
-                        Preprocessors.preprocessRequest(),
-                        Preprocessors.preprocessResponse(
-                                ResponseModifyingPreprocessors.replaceBinaryContent(),
-                                ResponseModifyingPreprocessors.limitJsonArrayLength(objectMapper),
-                                Preprocessors.prettyPrint())))
-                .apply(MockMvcRestDocumentation.documentationConfiguration(restDocumentation)
-                        .uris()
-                        .withScheme("https")
-                        .withHost("udhd.djbaek.com")
-                        .and().snippets()
-                        .withDefaults(CliDocumentation.curlRequest(),
-                                HttpDocumentation.httpRequest(),
-                                HttpDocumentation.httpResponse(),
-                                AutoDocumentation.requestFields(),
-                                AutoDocumentation.responseFields(),
-                                AutoDocumentation.pathParameters(),
-                                AutoDocumentation.requestParameters(),
-                                AutoDocumentation.description(),
-                                AutoDocumentation.methodAndPath(),
-                                AutoDocumentation.authorization("User access token required."),
-                                AutoDocumentation.sectionBuilder()
-                                        .snippetNames(
-                                                SnippetRegistry.AUTO_AUTHORIZATION,
-                                                SnippetRegistry.AUTO_PATH_PARAMETERS,
-                                                SnippetRegistry.AUTO_REQUEST_PARAMETERS,
-                                                SnippetRegistry.AUTO_REQUEST_FIELDS,
-                                                SnippetRegistry.HTTP_REQUEST,
-                                                SnippetRegistry.AUTO_RESPONSE_FIELDS,
-                                                SnippetRegistry.HTTP_RESPONSE)
-                                        .skipEmpty(true)
-                                        .build()))
-                .build();
-    }
-
-    protected RequestPostProcessor userToken() {
-        return (request) -> {
-            request.addHeader("Authorization", "Bearer <access-token>");
-            return documentAuthorization(request, "User access token required.");
-        };
-    }
-}
