+++ src/test/java/com/udhd/apiserver/web/FeedControllerTest.java
@@ -44,16 +44,90 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-
-public class FeedControllerTest extends ControllerTest{
+@ExtendWith(RestDocumentationExtension.class)
+@AutoConfigureRestDocs
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@SpringBootTest
+public class FeedControllerTest {
+    @Mock
+    private PhotoRepository photoRepository;
 
     @Autowired
-    protected UserController userController;
+    protected WebApplicationContext context;
 
+    @Autowired
+    protected ObjectMapper objectMapper;
+    @Autowired
+    protected UserController userController;
+    @Value("${userId}")
+    private String userId;
+    @Value("${feedId}")
+    private String feedId;
+
+    protected MockMvc mockMvc;
+
+    private MockedStatic<SecurityUtils> mockedSecurityUtils;
 
     private final PhotoOutlineDto mockPhotoOutlineDto = PhotoOutlineDto.builder()
             .photoId("456").thumbnailLink("http://link.com").build();
 
+    @BeforeAll
+    public void mockStaticSetup() {
+        mockedSecurityUtils = mockStatic(SecurityUtils.class);
+        given(SecurityUtils.getLoginUserId()).willReturn("612e4785662b04006f78157d");
+    }
+
+    @AfterAll
+    public void demockStaticSetup() {
+        mockedSecurityUtils.close();
+    }
+
+    @BeforeEach
+    public void setUp(RestDocumentationContextProvider restDocumentation) throws Exception {
+        this.mockMvc = MockMvcBuilders
+                .webAppContextSetup(context)
+                .alwaysDo(JacksonResultHandlers.prepareJackson(objectMapper))
+                .alwaysDo(MockMvcRestDocumentation.document("{class-name}/{method-name}",
+                        Preprocessors.preprocessRequest(),
+                        Preprocessors.preprocessResponse(
+                                ResponseModifyingPreprocessors.replaceBinaryContent(),
+                                ResponseModifyingPreprocessors.limitJsonArrayLength(objectMapper),
+                                Preprocessors.prettyPrint())))
+                .apply(MockMvcRestDocumentation.documentationConfiguration(restDocumentation)
+                        .uris()
+                        .withScheme("https")
+                        .withHost("udhd.djbaek.com")
+                        .and().snippets()
+                        .withDefaults(CliDocumentation.curlRequest(),
+                                HttpDocumentation.httpRequest(),
+                                HttpDocumentation.httpResponse(),
+                                AutoDocumentation.requestFields(),
+                                AutoDocumentation.responseFields(),
+                                AutoDocumentation.pathParameters(),
+                                AutoDocumentation.requestParameters(),
+                                AutoDocumentation.description(),
+                                AutoDocumentation.methodAndPath(),
+                                AutoDocumentation.authorization("User access token required."),
+                                AutoDocumentation.sectionBuilder()
+                                        .snippetNames(
+                                                SnippetRegistry.AUTO_AUTHORIZATION,
+                                                SnippetRegistry.AUTO_PATH_PARAMETERS,
+                                                SnippetRegistry.AUTO_REQUEST_PARAMETERS,
+                                                SnippetRegistry.AUTO_REQUEST_FIELDS,
+                                                SnippetRegistry.HTTP_REQUEST,
+                                                SnippetRegistry.AUTO_RESPONSE_FIELDS,
+                                                SnippetRegistry.HTTP_RESPONSE)
+                                        .skipEmpty(true)
+                                        .build()))
+                .build();
+    }
+
+    protected RequestPostProcessor userToken() {
+        return (request) -> {
+            request.addHeader("Authorization", "Bearer <access-token>");
+            return documentAuthorization(request, "User access token required.");
+        };
+    }
     @Test
     @DisplayName("Feed Like")
     void likeFeed() throws Exception {
